{"types":"C","runs":16}

+++ problem17.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class problem17 extends ConsoleProgram{
     public void run(){
-    	double num = readInt("ENTER THE NUMBER: ");
+    	double num = readDouble("ENTER THE NUMBER: ");
     	int ans1 = 0;
         for(ans1 = 0; ans1 < num; ans1++){
